requirements.txt
streamlit
pandas
numpy
scikit-learn
google-genai